пред релиз бета теста
закончены модули и шаблоны
не разбит на модули ввод/вывод, вводится Tkinter
