+++ default/lattice_default
@@ -1,0 +1,3 @@
+
+% -------------------- LATTICE --------------------
+
