+++ default/materials_default
@@ -1,3 +1,6 @@
+
+% -------------------- MATERIALS --------------------
+
 % --- Zr-Nb cladding and shroud tube:
 
 mat clad    -6.55000 rgb 100 100 100
