+++ default/pins_default
@@ -1,3 +1,6 @@
+
+% -------------------- PINS --------------------
+
 % --- Central tube:
 
 pin c
