+++ default/surfaces_default
@@ -1,0 +1,13 @@
+
+% -------------------- SURFACES --------------------
+
+surf tvs_hex	        hexyc	0.0  0.0  11.6625  	% Shroud tube inner radius
+
+surf tvel_bottom        pz 	    -188.5	            %tvel bottom
+surf tvel_top           pz  	210.7	            %tvel top
+
+surf tvs_bottom         pz 	    -215.2              %tvs bottom
+surf tvs_top            pz  	249.6	            %tvs top
+
+surf fuel_top           pz  	 184.0 	            %fuel top
+surf fuel_bottom        pz 	    -184.0	            %fuel bottom
