+++ modules/file_gen/debugger.py
@@ -1,0 +1,5 @@
+
+def debugger(message):
+    print_debug_data = False
+    if print_debug_data == True:
+        print(message)
