+++ modules/interface/interface.py
@@ -1,0 +1,7 @@
+from tkinter import *
+
+def interface():
+    root = Tk()
+    root.title('My App') 
+    root.mainloop()
+    return True
