+++ templates/file_start_template
@@ -1,0 +1,5 @@
+% variant: {name}
+
+% --------------------------------------\ U49G6 Assembly --------------------------------------
+
+
