+++ templates/lattice_template
@@ -1,0 +1,27 @@
+lat l{i} 2  0.0 0.0 23 23 1.275
+
+{w} {w} {w} {w} {w} {w} {w} {w} {w} {w} {w} {w} {w} {w} {w} {w} {w} {w} {w} {w} {w} {w} {w} 
+ {w} {w} {w} {w} {w} {w} {w} {w} {w} {w} {w} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {w} 
+  {w} {w} {w} {w} {w} {w} {w} {w} {w} {w} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {w} 
+   {w} {w} {w} {w} {w} {w} {w} {w} {w} {i} {i} {g} {i} {i} {i} {i} {i} {i} {i} {g} {i} {i} {w} 
+    {w} {w} {w} {w} {w} {w} {w} {w} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {w} 
+     {w} {w} {w} {w} {w} {w} {w} {i} {i} {i} {i} {i} {i} {i} {c} {i} {i} {i} {i} {i} {i} {i} {w} 
+      {w} {w} {w} {w} {w} {w} {i} {i} {i} {i} {i} {c} {i} {i} {i} {i} {c} {i} {i} {i} {i} {i} {w} 
+       {w} {w} {w} {w} {w} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {w} 
+        {w} {w} {w} {w} {i} {i} {i} {i} {c} {i} {i} {i} {c} {i} {i} {i} {i} {c} {i} {i} {i} {i} {w} 
+         {w} {w} {w} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {c} {i} {i} {i} {i} {i} {i} {i} {w} 
+          {w} {w} {i} {i} {i} {i} {i} {i} {i} {c} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {w} 
+           {w} {i} {i} {g} {i} {i} {c} {i} {i} {i} {i} {c} {i} {i} {i} {i} {c} {i} {i} {g} {i} {i} {w} 
+            {w} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {c} {i} {i} {i} {i} {i} {i} {i} {w} {w} 
+             {w} {i} {i} {i} {i} {i} {i} {i} {c} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {w} {w} {w}              
+              {w} {i} {i} {i} {i} {c} {i} {i} {i} {i} {c} {i} {i} {i} {c} {i} {i} {i} {i} {w} {w} {w} {w}            	
+               {w} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {w} {w} {w} {w} {w}             	 
+                {w} {i} {i} {i} {i} {i} {c} {i} {i} {i} {i} {c} {i} {i} {i} {i} {i} {w} {w} {w} {w} {w} {w}               	  
+                 {w} {i} {i} {i} {i} {i} {i} {i} {c} {i} {i} {i} {i} {i} {i} {i} {w} {w} {w} {w} {w} {w} {w}                	  
+                  {w} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {w} {w} {w} {w} {w} {w} {w} {w}               	    
+                   {w} {i} {i} {g} {i} {i} {i} {i} {i} {i} {i} {g} {i} {i} {w} {w} {w} {w} {w} {w} {w} {w} {w}               	    
+                    {w} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {w} {w} {w} {w} {w} {w} {w} {w} {w} {w}               	     
+                     {w} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {i} {w} {w} {w} {w} {w} {w} {w} {w} {w} {w} {w}             	     
+                      {w} {w} {w} {w} {w} {w} {w} {w} {w} {w} {w} {w} {w} {w} {w} {w} {w} {w} {w} {w} {w} {w} {w}
+
+
