+++ templates/surface_template
@@ -1,0 +1,1 @@
+surf fuel_surf{iter}    pz  	{height}	        %fuel top
